// apps/inventory/src/app/remote-entry/model/mastercategorytype.interface.ts
  export interface CATEGORYTYPES {
    id:number;
    code: string;
    description: string;
  }

  export interface CATEGORYTYPE{
    id: number;
    code: string;
    description: string;
    createdBy: number;
    createdOn: string;
    activeFlag: number;
    avgStockQuantity: number;
  }
  

  

  